Split Android local repositories
